Move `TypeNodePass` from `fuzzyc2cpg` to `semanticcpg`
