+++ fuzzyc2cpg/src/test/scala/io/shiftleft/fuzzyc2cpg/passes/TypeNodePassTests.scala
@@ -4,6 +4,7 @@ import better.files.File
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.passes.IntervalKeyPool
 import io.shiftleft.semanticcpg.language._
+import io.shiftleft.semanticcpg.passes.typenodes.TypeNodePass
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
