+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/passes/typenodes/TypeNodePass.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.fuzzyc2cpg.passes
+package io.shiftleft.semanticcpg.passes.typenodes
 
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes
